operf, remove support to report multiplexing.
The multiplexing reporting doesn't work correctly when profiling
multi-threaded apps or apps that do fork/exec.  The detection of
multiplexing doesn't work when processes migrate between CPUs.
The event is enabled on all CPUs.  The running time stops when the
event migrates to another CPU however, the enabled time does not stop as it
is enabled on each CPU.  The issue is that the running time across CPUs
doesn't add up to the enabled time because of the running time is not
increasing while the process is being migrated.  This results in the running
time being less then the enabled time.  There is no way to detect if the
reason the running time is less then the enabled time was do to migration
or due to multiplexing.

The support is being removed so that the operf tool is not incorrectly
flagging events for multiplexing.

Signed-off-by: Carl Love <cel@us.ibm.com>