Fix recent regression involving unit mask values of '0'
A patch that I wrote and committed on Jan 23, 2013 broke
unit mask handling for mask values of '0'.  As far as I
can tell, i386/core_2 is the only cpu type that has unit
masks with values of '0'.  So, for example, the following
command will fail as shown below:

   operf -e L2_M_LINES_OUT:100000:0 ls
   Invalid unit mask 0x0 for event L2_M_LINES_OUT
   Error retrieving info for event L2_M_LINES_OUT:100000:0

This patch fixes that problem.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>