Fix size problem of oprof_start GUI
On oprof_start startup, "unit mask" sub-window sometimes overlaps "event list"
one ( confusing appearance ).  On some systems, the symptom is the unit mask
sub-window is too narrow to fit the entire UM description, thus requiring the
user to manually re-size the whole GUI window manually.

To see this problem, select a long event name that also has some long unit
mask descriptions that would entail auto-resizing of the windows. Then save
and quit, and restart the gui.  The windows may be overlapped or may not be
sized correctly to accommodate the long event name and/or long unit mask
description.

This problem occurs if oprof_start is built with Qt4 / Qt3Support.

This patch fixes this.

Signed-off-by: Gilles Allard <gilles.allard21@orange.fr>