Add EXTRA_NONE flag for use with unit masks empty "extra:" field
Add a dummy extra flag that is set when any extra: (including an empty one)
field is present in a unit mask line.

This can be used to express that the unit mask has a unique name as the
first word of the description.

This is mapped to the x86 enabled flag, as we have to pass it around
as part of the event mask.

Signed-off-by: Andi Kleen <ak@linux.intel.com>