Add empty extra: lines to every Intel event with a unique first word
Add empty extra: lines to every Intel event with a unique first word
in the description. This can then be used to specify the unit mask
symbolically.

Haswell already had the empty extra masks.
Core 2 did not have unique words for everything. I only add it there
to the events which had.

Signed-off-by: Andi Kleen <ak@linux.intel.com>