Add 'check_count' parameter to parse_events function
For profiling, oprofile enforces a 'count' value to be
included in the event specification passed by a user.
It is the parse_events() function in libop/op_parse_event.c
that handles that checking.  But for counting events
versus profiling, a 'count' value is meaningless, so this
patch adds support for parse_events() to parse an event
spec from either the profiling tools or the new counting
tool.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>