ophelp does not always detect duplicate numerical unit mask values
Some named unit masks have hex values that are duplicates of others in the
same UM entry, thus requiring the use of the name to disambiguate them.
Specifying one of the hex values that are duplicated should result in an
error -- but there are cases where it doesn't.  For example:
 Here's a properly working example on Sandybridge:
   [mpjohn@oc1757000783 test-stuff]$ ophelp --check-events int_misc:2000000:0x3
   Unit mask (0x3) is non unique.
   Please specify the unit mask using the first word of the description

 and here's a failing example:
   [mpjohn@oc1757000783 test-stuff]$ ophelp --check-events l1d_pend_miss:2000000:0x1
   2

This patch fixes the problem.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>