OProfile, fix the units for the reported CPU frequency
The freqency of the processor is found by function op_cpu_frequency() in
libutil/op_cpufreq.c by either checking for the frequency in the file
/sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq or from /proc/cpuinfo.

Most of the Intel processors get the frequency from cpuinfo_max_freq.  The
frequency stored in this file is stored in units of KHz for the Intel
processors I can check.  I do not know what other architectures store the
CPU frequency in this file.  When this value is printed, the value is printed
by the routine describe_cpu() in libpp/op_header.cpp the assumption is the
freqency is assumed to be MHz.  For example, the following is what is printed
by my laptop after running opreport

  Using /home/carll/oprofile_data/samples/ for samples directory.
  CPU: Core 2, speed 2.534e+06 MHz (estimated)
  Counted CPU_CLK_UNHALTED events (Clock cycles when not halted) with a unit ma\
sk of 0x00 (Unhalted core cycles) count 100000

Note the speed is stated as 2.534e+06 MHz that is Mega Mega Hz or Tera Hz which
is a factor of 1000 high.

The fix is to have the function op_cpu_freq_sys_devices() in
libutil/op_cpufreq.c to adjust the frequency read from the file
/sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq by dividing by 1000
to return the frequency in units of MHz.  The following patch implements this
fix for reporting the estimated processor frequency.

Signed-off-by: Carl Love <cel@us.ibm.com>