Fix operf/ocount default unit mask selection
Many events (particularly in the x86* architectures)
require a unit mask value to specify the exact event
type.  For such events, a default unit mask value
is assigned.  When a user runs operf, ocount, or
opcontrol and specifies such an event but does not
specify a unit mask, the default unit mask should be
selected and used by the tool.  A bug was discovered
with operf and ocount where the unit mask value in
this situation was being set to '0' instead of the
default unit mask value.  This patch fixes the bug.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>