Add more helpful info about dealing with lost samples
When operf detects that more than a certain percentage of
samples were lost, it displays a warning message when it
stops.  This patch adds to that message a suggestion to
lower the sampling rate.  This patch also updates the
operf man page with information on how to control the
sampling rate.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>