Fix regression in IBM POWER8 running in POWER7 compat mode
A commit made on Dec 17, 2013 ("Allow all native events for IBM POWER8
in POWER7 compat mode) broke support for POWER8 in POWER7 compat mode.
Instead, oprofile attempts to treat it as a normal POWER7 processor,
which is not correct.  A user reported the following error when
running operf with the default CYCLES event:

   terminate called after throwing an instance of 'std::runtime_error'
     what():  libpfm cannot find event code for CYCLES; cannot continue
   Aborted

This patch fixes this problem.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>