Enhance ocount to support millisecond time intervals
The ocount tool currently has a --time-interval option that
allows users to print a snapshot of event counts every n seconds,
where 'n' is an integer value equal to or greater than 1.
A user requested a change to allow time intervals of 100 ms.
The user stated that he uses an event counting tool on a
non-Linux operating system, and that tool had the option of
collecting counts every 100 ms. Now, he wanted the same ability
on Linux so he could do direct comparison of data. The
original choice of one second granularity for the time interval
option was arbitrary, so the request (based on a real world
use case) seemed reasonable.

This patch changes ocount's '--time-interval' option to accept
values in units of milliseconds instead of seconds. At this
time, the implementation supports only 100 ms granularity, so
time intervals passed will be rounded to the nearest 100 ms.

NOTE: Current users of ocount who make use of the '--time-interval'
option will see a change in behavior. I don't believe this is
a serious compatibility issue since the timestamp printed along
with the event counts makes it very clear what the actual
time interval is, and the user can check the ocount man page
for clarification.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>