Plug timing hole between JIT agent and opjitconv that can corrupt dump file
There's a window of time where oprofile's Java JIT agent library may be
writing to a JIT dump file at a time when the opjitconv is simultaneously
making a copy of it for purposes of processing its records to create an
ELF file.  It's possible that opjitconv may make a copy of the file when
it's not in a coherent state -- i.e., the JIT agent may be in the process
of performing multiple writes for a record. There is no OS-level locking
of the file, so neither of the processes involved (the JIT agent and
opjitconv) are aware the other one is accessing the file.

No actual bugs have been reported due to this problem, so this patch is
a pre-emptive fix. Symptoms of this problem occurring would include
opjitconv not processing all JIT dump files, and operf displaying the message
    JIT dump processing exited abnormally: 1
Passing "--verbose=debug" to operf to collect detailed information on the JIT
conversion process would show the following error message:

    record past end of file
    JIT convert error -1

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>