Remove opreport warnings for /no-vmlinux, [vdso], [hypervisor_bucket] not found
When running opreport -l, the following warning messages may appear:

warning: /no-vmlinux could not be found.
warning: [vdso] (tgid:31882 range:0x7fff249ae000-0x7fff249aefff) could not be found.
warning: [hypervisor_bucket] (tgid:24654 range:0x994-0x17a09c0) could not be found

For the '/no-vmlinux' case, it's not a real file -- it's just a bucket
we use to assign kernel samples to.  For the '[vdso]' case, this reflects
samples in the vDSO (virtual dynamic shared object), which is a small shared
library that the kernel automatically maps into the address space of all
user-space applications.  The '[hypervisor_bucket]' message can happen
on virtualized systems with a hypervisor (e.g., PowerVM) because there is
no physical binary file to attribute hypervisor samples to.

The warnings are annoying and unnecessary, so they are removed via this
patch. This patch also adds a description of "no-vmlinux" to the operf man
page.  And for '[vdso]', users can easily google for a description.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>