Make op_bfd::get_kallsym_symbols() 32/64-bit agnostic
The size of %lx is 32 bits on 32-bit machine and 64-bits on 64-bit
machines.  However, bfd_vma is always a 64-bit value regardless
whether the processor is a 32-bit or 64-bit architecture.  The
get_kallsym_symbols() used the %lx which only matched up on the 64-bit
machines.  This changes the code to always read things into a
"unsigned long long" variable and then copy that value into the
bfd_vma variable to allow the code to compile on both 32-bit and
64-bit machine