Bug #266: exclude/include files option doesn't work for opannotate -a
The opannotate man page did not make it clear at all that the --exclude-file
and --include-file options apply *only* when specifying "--source"
annotation.  This patch updates the man page to clarify the behavior.
It also adds a check in opannotate to disallow specifying either of these
options when "--assembly" is specified.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>