operf log may over-report "sample address not in expected range for domain"
When find_mapping_for_sample() is called, only the sample address is passed.
However, in certain situations, the address alone is not enough to find the
proper matching operf_mmap object.  For example, hypervisor addresses can
overlap with userspace addresses where the vdso gets mapped in a 32-bit
process on the ppc64 architecture.  If an operf_mmap object has already been
created with a particular memory address range and then a sample address
from VDSO in a 32-bit process needs to be processed, we may incorrectly match
that address with the hypervisor operf_mmap object.  With this patch, we now
will pass a flag ('hypervisor_sample') and ensure that we only return an
operf_mmap object if the address is a match AND the flag value matches with
operf_mmap.is_hypervisor.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>