Minor error-handling fixes needed for ocount
The main fix in this patch is to properly handle the case
where perf_event_open fails after we have already started
the app for which the user requested to count events.
Prior to this fix, ocount would simply exit, leaving the
app to run to completion.  This patch causes ocount
to kill the app if we encounter a perf_event_open error.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>