opreport: header timestamps are different for kallsyms file
The time stamp for kallsyms can be different because it is not a real
file.  Hence, when there are samples from the kallsyms file and the
following conditions are met you get an error about the time stamps
not matching.

  - operf was run with the '--separate-thread' option
  - operf was run either as root or as normal user where
    /proc/sys/kernel/kptr_restrict is set to 0
  - The application being profiled is a multi-threaded app that
    executes both pthread_create and fork[2]

This patch fixes the issue by assigning the time stamp of zero when
the source file is kallsyms.

Signed-off-by: Carl Love <cel@us.ibm.com>