Fix sample data pipe partial read handling
Unless the user specifies the "--lazy-conversion" option with operf,
there will be separate "record" and "convert" processes started
by operf. The record process reads the sample data from the kernel
(via mmap'ed memory) and sends that data to the convert process
over a pipe. The convert process takes the raw sample data
and converts it to oprofile sample files. In the procedure for
reading the pipe, the convert process does two reads: one read to
obtain the sample data event header (which will give the size of
the whole sample data event), and a second read to get the rest
of the event record. For both of these reads, there is code to handle
return values that imply something other than a successful read of the
record. If the value returned is > 0 but less than the expected
length, we go back and do the read *again* to get the rest of the
record. However, in the second read (to obtain the bulk of the event
record), we were doing a "goto" to the wrong place -- we were going
back to the location of the first read.

This problem, in actual practice, is not likely to occur, since the
record process should normally be able to write an entire event
record to the pipe wihout interruption. But if it does occur, the
convert process will probably fail with a message like the following:

Event num -1 for id <x> is invalid. Sample data appears to be corrupted.

In fact, about a year ago, a user had reported such a problem. It was
intermittent, and was not reproducible except on that user's particular
system, which was a very large server system. He was doing a system-wide
profile of a Java benchmark. By lowering the sampling rate, the problem
went away. At the time, I was not able to find the problem, but I suspect
that the problem described above that's addressed by this patch is very
likely the cause of his problem, too. I've contacted this user, but so
far, he's not been able to reproduce the original problem in order to
effectively test this patch. Nevertheless, I'm confident that this fix
is a correct one and should be done.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>